pi: shrink to re-scoped surface — pure-function tables, hand-executed proof at validation
